Refactoring for cleaner component structure, StudyInstanceUIDHandler introduced
