+++ shanoir-ng-datasets/src/main/java/org/shanoir/ng/dicom/web/service/DICOMWebService.java
@@ -1,7 +1,5 @@
 package org.shanoir.ng.dicom.web.service;
 
-import java.io.InputStream;
-
 import javax.annotation.PostConstruct;
 
 import org.apache.http.HttpEntity;
